system-monitor-pro/requirements.txt
psutil>=5.9.0,<6.0.0
paho-mqtt>=2.0.0,<3.0.0
